Integrated the camera system more
Added the rest of the camera options into the system, also added texture support

+++ Source/Core/Assets/AssetId.hpp
@@ -1,6 +1,6 @@
 #ifndef ASSETID_HPP
 #define ASSETID_HPP
 
-using AssetID = long long unsigned;
+using AssetID = unsigned;
 
 #endif
